+++ Sources/Controllers/ALKConversationViewController+TableView.swift
@@ -48,6 +48,7 @@ extension ALKConversationViewController: UITableViewDelegate, UITableViewDataSou
         case .text, .html, .email:
             if !configuration.isLinkPreviewDisabled, message.messageType == .text, ALKLinkPreviewManager.extractURLAndAddInCache(from: message.message, identifier: message.identifier) != nil {
                 var cell = ALKLinkPreviewBaseCell()
+                cell.menuOptionsToShow = configuration.messageMenuOptions
                 if message.isMyMessage {
                     cell = tableView.dequeueReusableCell(forIndexPath: indexPath) as ALKMyLinkPreviewCell
                     cell.showReport = false
@@ -70,11 +71,13 @@ extension ALKConversationViewController: UITableViewDelegate, UITableViewDataSou
             } else {
                 if message.isMyMessage {
                     let cell = tableView.dequeueReusableCell(forIndexPath: indexPath) as ALKMyMessageCell
+                    cell.menuOptionsToShow = configuration.messageMenuOptions
                     cell.showReport = false
                     necessarySetupForMessageCell(cell: cell, message: message)
                     return cell
                 } else {
                     let cell = tableView.dequeueReusableCell(forIndexPath: indexPath) as ALKFriendMessageCell
+                    cell.menuOptionsToShow = configuration.messageMenuOptions
                     cell.showReport = configuration.isReportMessageEnabled
                     cell.avatarTapped = { [weak self] in
                         guard let currentModel = cell.viewModel else { return }
@@ -93,6 +96,7 @@ extension ALKConversationViewController: UITableViewDelegate, UITableViewDataSou
                     // Set the value to nil so that previous image gets removed before reuse
                     cell.setLocalizedStringFileName(configuration.localizedStringFileName)
                     cell.photoView.image = nil
+                    cell.menuOptionsToShow = configuration.messageMenuOptions
                     cell.showReport = false
                     cell.update(viewModel: message)
                     cell.uploadTapped = { [weak self]
@@ -115,6 +119,7 @@ extension ALKConversationViewController: UITableViewDelegate, UITableViewDataSou
 
                 } else {
                     let cell: ALKMyPhotoLandscapeCell = tableView.dequeueReusableCell(forIndexPath: indexPath)
+                    cell.menuOptionsToShow = configuration.messageMenuOptions
                     cell.showReport = false
                     cell.setLocalizedStringFileName(configuration.localizedStringFileName)
                     cell.update(viewModel: message)
@@ -132,6 +137,7 @@ extension ALKConversationViewController: UITableViewDelegate, UITableViewDataSou
                 if message.ratio < 1 {
                     let cell: ALKFriendPhotoPortalCell = tableView.dequeueReusableCell(forIndexPath: indexPath)
                     cell.setLocalizedStringFileName(configuration.localizedStringFileName)
+                    cell.menuOptionsToShow = configuration.messageMenuOptions
                     cell.showReport = configuration.isReportMessageEnabled
                     cell.update(viewModel: message)
                     cell.downloadTapped = { [weak self]
@@ -149,6 +155,7 @@ extension ALKConversationViewController: UITableViewDelegate, UITableViewDataSou
 
                 } else {
                     let cell: ALKFriendPhotoLandscapeCell = tableView.dequeueReusableCell(forIndexPath: indexPath)
+                    cell.menuOptionsToShow = configuration.messageMenuOptions
                     cell.showReport = configuration.isReportMessageEnabled
                     cell.menuAction = { [weak self] action in
                         self?.menuItemSelected(action: action, message: message)
@@ -165,6 +172,7 @@ extension ALKConversationViewController: UITableViewDelegate, UITableViewDataSou
 
             if message.isMyMessage {
                 let cell: ALKMyVoiceCell = tableView.dequeueReusableCell(forIndexPath: indexPath)
+                cell.menuOptionsToShow = configuration.messageMenuOptions
                 cell.showReport = false
                 cell.setLocalizedStringFileName(configuration.localizedStringFileName)
                 cell.downloadTapped = { [weak self] _ in
@@ -178,6 +186,7 @@ extension ALKConversationViewController: UITableViewDelegate, UITableViewDataSou
                 return cell
             } else {
                 let cell: ALKFriendVoiceCell = tableView.dequeueReusableCell(forIndexPath: indexPath)
+                cell.menuOptionsToShow = configuration.messageMenuOptions
                 cell.showReport = configuration.isReportMessageEnabled
                 cell.setLocalizedStringFileName(configuration.localizedStringFileName)
                 cell.downloadTapped = { [weak self] _ in
@@ -197,6 +206,7 @@ extension ALKConversationViewController: UITableViewDelegate, UITableViewDataSou
         case .location:
             if message.isMyMessage {
                 let cell: ALKMyLocationCell = tableView.dequeueReusableCell(forIndexPath: indexPath)
+                cell.menuOptionsToShow = configuration.messageMenuOptions
                 cell.showReport = false
                 cell.setLocalizedStringFileName(configuration.localizedStringFileName)
                 cell.update(viewModel: message)
@@ -209,6 +219,7 @@ extension ALKConversationViewController: UITableViewDelegate, UITableViewDataSou
             } else {
                 let cell: ALKFriendLocationCell = tableView.dequeueReusableCell(forIndexPath: indexPath)
                 cell.setLocalizedStringFileName(configuration.localizedStringFileName)
+                cell.menuOptionsToShow = configuration.messageMenuOptions
                 cell.showReport = configuration.isReportMessageEnabled
                 cell.update(viewModel: message)
                 cell.setDelegate(locDelegate: self)
@@ -229,6 +240,7 @@ extension ALKConversationViewController: UITableViewDelegate, UITableViewDataSou
         case .video:
             if message.isMyMessage {
                 let cell: ALKMyVideoCell = tableView.dequeueReusableCell(forIndexPath: indexPath)
+                cell.menuOptionsToShow = configuration.messageMenuOptions
                 cell.showReport = false
                 cell.setLocalizedStringFileName(configuration.localizedStringFileName)
                 cell.update(viewModel: message)
@@ -252,6 +264,7 @@ extension ALKConversationViewController: UITableViewDelegate, UITableViewDataSou
             } else {
                 let cell: ALKFriendVideoCell = tableView.dequeueReusableCell(forIndexPath: indexPath)
                 cell.setLocalizedStringFileName(configuration.localizedStringFileName)
+                cell.menuOptionsToShow = configuration.messageMenuOptions
                 cell.showReport = configuration.isReportMessageEnabled
                 cell.update(viewModel: message)
                 cell.downloadTapped = { [weak self]
@@ -271,6 +284,7 @@ extension ALKConversationViewController: UITableViewDelegate, UITableViewDataSou
 
             if message.isMyMessage {
                 let cell = tableView.dequeueReusableCell(forIndexPath: indexPath) as ALKMyGenericCardMessageCell
+                cell.menuOptionsToShow = configuration.messageMenuOptions
                 cell.showReport = false
                 cell.setLocalizedStringFileName(configuration.localizedStringFileName)
                 cell.register(cell: ALKGenericCardCell.self)
@@ -281,6 +295,7 @@ extension ALKConversationViewController: UITableViewDelegate, UITableViewDataSou
                 return cell
             } else {
                 let cell = tableView.dequeueReusableCell(forIndexPath: indexPath) as ALKFriendGenericCardMessageCell
+                cell.menuOptionsToShow = configuration.messageMenuOptions
                 cell.showReport = configuration.isReportMessageEnabled
                 cell.setLocalizedStringFileName(configuration.localizedStringFileName)
                 cell.register(cell: ALKGenericCardCell.self)
@@ -382,6 +397,7 @@ extension ALKConversationViewController: UITableViewDelegate, UITableViewDataSou
             if message.isMyMessage {
                 let cell: ALKMyDocumentCell = tableView.dequeueReusableCell(forIndexPath: indexPath)
                 cell.setLocalizedStringFileName(configuration.localizedStringFileName)
+                cell.menuOptionsToShow = configuration.messageMenuOptions
                 cell.showReport = false
                 cell.update(viewModel: message)
                 cell.update(chatBar: chatBar)
@@ -402,6 +418,7 @@ extension ALKConversationViewController: UITableViewDelegate, UITableViewDataSou
                 return cell
             } else {
                 let cell: ALKFriendDocumentCell = tableView.dequeueReusableCell(forIndexPath: indexPath)
+                cell.menuOptionsToShow = configuration.messageMenuOptions
                 cell.showReport = configuration.isReportMessageEnabled
                 cell.setLocalizedStringFileName(configuration.localizedStringFileName)
                 cell.update(viewModel: message)