+++ Sources/Views/ALKVoiceCell.swift
@@ -284,7 +284,7 @@ class ALKVoiceCell: ALKChatBaseCell<ALKMessageViewModel>,
     }
 
     func menuReport(_: Any) {
-        menuAction?(.reportMessage)
+        menuAction?(.report)
     }
 }